+++ src/navigation/AppNavigator.tsx
@@ -4,23 +4,23 @@ import {
   createNativeStackNavigator,
 } from "@react-navigation/native-stack";
 import { TabNavigator } from "./TabNavigator";
-import WelcomeScreen from "@/pages/onboarding/welcome/Welcome";
-import ConnectionDetailScreen from "@/pages/default/connections/ConnectionDetail";
-import ReviewConnectionScreen from "@/pages/default/connections/ReviewConnection";
-import AddCredentialsScreen from "@/pages/default/credentials/AddCredentials";
-import CredentialDetailScreen from "@/pages/default/credentials/CredentialDetail";
-import ProfileDetailScreen from "@/pages/default/profiles/ProfileDetail";
-import CreatePassphraseScreen from "@/pages/onboarding/create/CreatePassphrase";
-import CreateProfilesScreen from "@/pages/onboarding/create/CreateProfiles";
-import CreateWalletScreen from "@/pages/onboarding/create/CreateWallet";
-import AddProfileScreen from "@/pages/default/profiles/AddProfile";
-import AddCredentialDetailScreen from "@/pages/default/credentials/AddCredentialDetail";
-import AddCredentialOptionsScreen from "@/pages/default/credentials/AddCredentialOptions";
+import WelcomeScreen from "@/screens/onboarding/welcome/Welcome";
+import ConnectionDetailScreen from "@/screens/default/connections/ConnectionDetailScreen";
+import ReviewConnectionScreen from "@/screens/default/connections/ReviewConnectionScreen";
+import AddCredentialsScreen from "@/screens/default/credentials/AddCredentialsScreen";
+import CredentialDetailScreen from "@/screens/default/credentials/CredentialDetailScreen";
+import ProfileDetailScreen from "@/screens/default/profiles/ProfileDetailScreen";
+import CreatePassphraseScreen from "@/screens/onboarding/create/CreatePassphraseScreen";
+import CreateProfilesScreen from "@/screens/onboarding/create/CreateProfilesScreen";
+import CreateWalletScreen from "@/screens/onboarding/create/CreateWalletScreen";
+import AddProfileScreen from "@/screens/default/profiles/AddProfileScreen";
+import AddCredentialDetailScreen from "@/screens/default/credentials/AddCredentialDetailScreen";
+import AddCredentialOptionsScreen from "@/screens/default/credentials/AddCredentialOptionsScreen";
 import { IdentityAgentManager } from "@/features/identity/IdentityAgentManager";
-import LoadingScreen from "@/pages/Loading";
-import EnterPassphraseScreen from "@/pages/default/passphrase/EnterPassphrase";
-import ConnectionRequestScreen from "@/pages/default/connections/ConnectionRequest";
-import AddConnectionScreen from "@/pages/default/connections/AddConnectionScreen";
+import LoadingScreen from "@/components/Loading";
+import EnterPassphraseScreen from "@/screens/default/passphrase/EnterPassphraseScreen";
+import ConnectionRequestScreen from "@/screens/default/connections/ConnectionRequestScreen";
+import AddConnectionScreen from "@/screens/default/connections/AddConnectionScreen";
 import type { AppNavigatorInterface } from "@/types/navigation";
 
 const Stack = createNativeStackNavigator<AppNavigatorInterface>();
