+++ src/screens/default/connections/ConnectionDetailScreen.tsx
@@ -1,16 +1,16 @@
 import React, { useState } from "react";
 import { Text, View } from "react-native";
-import { MenuPageLayout } from "../MenuPageLayout";
+import { MenuPageLayout } from "../../../components/MenuPageLayout";
 import { LabelValueItem } from "@/components/LabelValue";
 import { formatDID } from "@/util/formatters";
 import { Layouts } from "@/theme/layouts";
 import { Typography } from "@/theme/typography";
-import { Tappable } from "../Tappable";
+import { Tappable } from "../../../components/Tappable";
 import { BadgeNames } from "@/components/Item";
-import { mockConnections } from "@/services/mocks";
 import { profilesAtom } from "@/features/profile/atoms";
 import { For } from "@legendapp/state/react";
 import { AppNavigatorProps } from "@/types/navigation";
+import { mockConnections } from "@/features/connect/mocks";
 
 const tabLabels = ["About", "Connections", "Activity"];
