+++ src/screens/default/credentials/AddCredentialsScreen.tsx
@@ -3,11 +3,11 @@ import { Typography } from "@/theme/typography";
 import { For } from "@legendapp/state/react";
 import React from "react";
 import { SafeAreaView, ScrollView, Text, View } from "react-native";
-import { Tappable } from "../Tappable";
+import { Tappable } from "../../../components/Tappable";
 import { observable } from "@legendapp/state";
-import { mockCredentials } from "@/services/mocks";
 import type { MockCredential } from "@/types/models";
 import { AppNavigatorProps } from "@/types/navigation";
+import { mockCredentials } from "@/features/credentials/mocks";
 
 type Props = AppNavigatorProps<"AddCredentialsScreen">;
 const AddCredentialsScreen = ({ navigation }: Props) => {
