+++ src/screens/default/profiles/ProfilesScreen.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { ParentPageLayout } from "@/pages/default/ParentPageLayout";
-import { Tappable } from "@/pages/default/Tappable";
+import { ParentPageLayout } from "@/components/ParentPageLayout";
+import { Tappable } from "@/components/Tappable";
 import { ScrollView, View } from "react-native";
 import { Button } from "@/components/Button";
 import { FlexLayouts } from "@/theme/layouts";
 import { formatDID } from "@/util/formatters";
 import { TabNavigatorProps } from "@/types/navigation";
 import type { ManagedIdentity } from "@web5/agent";
-import LoadingScreen from "@/pages/Loading";
+import LoadingScreen from "@/components/Loading";
 import { useProfiles } from "@/features/profile/hooks";
 import { useIdentityList } from "@/features/identity/hooks";
