+++ src/screens/onboarding/create/CreatePassphraseScreen.tsx
