+++ src/screens/onboarding/create/CreateWalletScreen.tsx
@@ -4,7 +4,7 @@ import { FlexLayouts, Layouts } from "@/theme/layouts";
 import { Typography } from "@/theme/typography";
 import { AppNavigatorProps } from "@/types/navigation";
 import { IdentityAgentManager } from "@/features/identity/IdentityAgentManager";
-import { defaultIdentities } from "@/pages/onboarding/create/default_identities";
+import { defaultIdentities } from "@/features/identity/default_identities";
 import { Deeplink } from "@/features/deeplink/deeplink";
 import { BiometricLogin } from "@/features/biometrics/biometric-login";
